Separated views and tests into multiple file for different application modules.
