+++ api/admin.py
@@ -1,3 +1,13 @@
 from django.contrib import admin
 
-# Register your models here.
+from .models import User, Task, Swipe, Conversation, Message
+
+# We register our database models here so that
+# the admin interface can edit these database
+# tables. This is only used for development.
+
+admin.site.register(User)
+admin.site.register(Task)
+admin.site.register(Swipe)
+admin.site.register(Conversation)
+admin.site.register(Message)
